+++ auth_saml/__manifest__.py
@@ -2,30 +2,24 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
-    'name': 'Saml2 Authentication',
-    'version': '11.0.2.0.0',
-    'category': 'Tools',
-    'author': 'XCG Consulting, Odoo Community Association (OCA)',
-    'maintainer': 'XCG Consulting',
-    'website': 'http://odoo.consulting',
-    'license': 'AGPL-3',
-    'depends': [
-        'base_setup',
-        'web',
-        'auth_crypt',
+    "name": "SAML2 Authentication",
+    "version": "11.0.3.0.0",
+    "category": "Tools",
+    "author": "XCG Consulting, Odoo Community Association (OCA)",
+    "maintainer": "XCG Consulting",
+    "website": "http://odoo.consulting",
+    "license": "AGPL-3",
+    "depends": ["base_setup", "web", "auth_crypt"],
+    "data": [
+        "data/ir_config_parameter.xml",
+        "security/ir.model.access.csv",
+        "views/auth_saml.xml",
+        "views/base_settings.xml",
+        "views/res_users.xml",
     ],
-    'data': [
-        'data/auth_saml.xml',
-        'data/ir_config_parameter.xml',
-        'security/ir.model.access.csv',
-        'views/auth_saml.xml',
-        'views/base_settings.xml',
-        'views/res_users.xml',
-    ],
-    'installable': True,
-    'auto_install': False,
-    'external_dependencies': {
-        'python': ['lasso'],  # >= 2.6.0
-    },
-    'development_status': 'stable',
+    "demo": ["demo/auth_saml.xml"],
+    "installable": True,
+    "auto_install": False,
+    "external_dependencies": {"python": ["lasso"]},  # >= 2.6.0
+    "development_status": "stable",
 }